docs(breaking-changes): add table of contents
this commit updates `BREAKING_CHANGES.md` by adding a table of contents
to the document, as well as an executive summary. the goal of this
commit is to align with our sibling project, the Ionic Framework, it
terms of layout of the document.

keeping with that philosophy, existing header tags have been updated to
be one level 'smaller' to accomodate the new structure