fixups: getting parse5 to work with stencil
this commit performs a few mechanical changes to get parse5 to work:
- remove @types/parse5 per https://github.com/inikulin/parse5/releases/tag/v7.0.0
- fix a JSDoc parse5 typo (not required, but might as well while we're
  here)

all existing type imports were updated when no concrete implementation
was used. this was done for consistency, rather that an explicit need to
do so.

we no longer import `Attribute` from parse5. this type now sits in
`Token`'s namespace. as a result, we import `type Token` and access it
through `Token`.

the type `ParserOptions` no longer defaults to the type `TreeAdapter`
for its generic parameter. instead, it now accepts a type `T extends
TreeAdapterTypeMap`. `TreeAdapterTypeMap` previously existed in v6.0 of
parse5, but was not a part of the `TreeAdapter` definition. It has
changed like so:
```diff
- export interface TreeAdapterTypeMap {
+ export interface TreeAdapterTypeMap<Node = unknown, ParentNode = unknown, ChildNode = unknown, Document = unknown, DocumentFragment = unknown, Element = unknown, CommentNode = unknown, TextNode = unknown, Template = unknown, DocumentType = unknown> {
```
where `TreeAdapter` now uses `TreeAdapterMap`
```diff
- export interface TreeAdapter {
+ export interface TreeAdapter<T extends TreeAdapterTypeMap = TreeAdapterTypeMap> {
```
the key takeaway being that `TreeAdapter` defaults to a
`TreeAdapterTypeMap` for its typings of `Node`, `ParentNode`, etc. those
types default to `unknown`, which gives us backwards compatability with
v6, whose arguments all defaulted to `unknown`.

`ParserOptions` has had its `treeAdapter` field change to be of type:
```diff
- treeAdapter?: T;
+ treeAdapter?: TreeAdapter<T>;
```
aligning with the type hierachy explained above