chore(test): upgrade karma infrastructure (#3682)
this commit upgrades our karma infrastructure, bumping the versions of
all `karma-*` packages. this is being done not necessarily to keep the
packages up to date (although it is a benefit), but to test out if new packages
could have the positive side effect of helping with the stability of tests in our
continuous integration environment

this commit removes a test hack from 'delegates-focus' where a test
would return early if the userAgent wasn't 'Chrome'. this test has been
silently passing in Firefox before now, but with the new infrastructure
updates, it now fails