feat(docs-readme): add overview to readme (#3635)
this commit adds an 'Overview' section to the generated `README.md` file
for a component. this implementation is simple (and arguably naive)
intentionally. there is quite a bit that folks could do in the context
of a JSDoc. rather than attempting to capture every use case and
potentially overcomplicate our solution, we start simple and can iterate
from there.

this commit also adds JSDoc to a portion of the codebase that is related to
the generation of documentation