fix(compiler): fix determination of when to `super()` in convert-decorators.ts
This fixes an issue where we were inadvertently adding a `super()` call
to 'synthetic' constructors created in the `updateConstructor` function.

In particular, this could happen when a class declaration had an
`extends` keyword on it, like

```ts
class Foo implements Bar {
    foo = "bar";
}
```

In that case we _don't_ want to add a `super()` call to any constructor
we create, because the class has no superclass. So we need to tell that
situation apart from

```ts
class Foo extends Bar {
    foo = "bar";
}
```