chore(compiler): remove references to deprecated decorators (#3685)
We removed the decorators `@PropWillChange` and `@PropDidChange` quite a
while ago:

https://github.com/ionic-team/stencil/commit/fa2b400cf36696365487b7c2445ab096ee354e50

However, it looks like a few references to those decorators stuck around
in the `MEMBER_DECORATORS_TO_REMOVE` array in
src/compiler/transformers/decorators-to-static/decorators-constants.ts.

This just removes those! Given that the behavior for these was aleady
removed the impact should be minimal / nil.