chore(ci): add node 18
this commit increments the versions of node & npm `package.json#volta`.
this change sets a new baseline for the versions of npm and node that
will be used for the day-to-day development on the project.

to coincide with this change, node 18 has been added to the github
actions workflow.

in order for the build process to continue to work, the openssl legacy
provider has been enabled for our build process. without it, rollup will
fail to generate a build, as it relies on older hashing functions that
are only available via the legacy provider.