chore(compiler): update to non-deprecated `ts.factory` methods
This updates some more `ts.create` methods to use the equivalent under
`ts.factory`. Like in #3713, this started with selecting a file
(`src/compiler/transformers/component-native/native-connected-callback.ts`)
to 'branch out' from and then using project-wide find and replace to fix
all usages of the `ts.create*` methods used in that particular file.

That turned out to be:

- `ts.createExpressionStatement` -> `ts.factory.createExpressionStatement`
- `ts.createAssignment` -> `ts.factory.createAssignment`
- `ts.createPropertyAccess` -> `ts.factory.createPropertyAccessExpression`
- `ts.createThis` -> `ts.factory.createThis`
- `ts.createCall` -> `ts.factory.createCallExpression`
- `ts.createMethod` -> `ts.factory.createMethodDeclaration`
- `ts.createBlock` -> `ts.factory.createBlock`