chore(repo): remove duplicate file
this commit removes a duplicate file (note the file name ending). this
looks like it was created in 63595bc540db9aa85295f5a8dc8738beb2a37611 as
a part of removing the legacy compiler. it doesn't appear to be used
anywhere (luckily!), and the contents are idential to
`remove-static-meta-properties.ts` in this directory