chore(github): upgrade github actions dependencies
this commit bumps all third-party actions that are used by the stencil
repository to the latest and greatest versions. this is motivated by the
+100 warnings we started to receive regarding Node v12 support being
deprecated in ci. many of the dependencies that we update here had
sub-dependencies that were using/invoking these warnings, hence the
upgrade.

this commit also updates the main and browserstack workflows to use the
version of volta declared in `package.json`. this functionality was
introduced in a newer version of `setup-node`, and provides us the
benefit of a single source of truth for which version to run CI steps in
(rather than keeping package.json and ci files in sync)