fix(docs): avoid duplicating manual documentation (#3766)
this commit fixes a regression that was introduced in 2db4f4de62c9547ccafb1d382130fcf82fd9ebf4
(#3635). the original commit did not take into account the fact that a
user could have manually written text in their component's markdown
file. that content would be assigned to the value used to generated the
markdown file's "Overview" section, causing the content to be displayed
twice.

the contents of the `docs` field (i.e. the way that value is generated).
cannot be changed, as `JsonDocs` publicly exposes that field (via a
`JsonDocsComponent` value), and may be used by custom documentation
generators.

instead, this commit adds an optional 'overview' field to the
`JsonDocsComponent` type. the intent here is to avoid _most_
assignability errors by making this field optional. the value is
assigned when stencil generates its documentation data with the text
from the class component's jsdoc.

this commit adds additional testing coverage to the 'docs' field of
`JsonDocsComponent`