refactor(compiler): create nodes using factory methods (#3748)
this commit updates stencil to create nodes in the typescript syntax
tree using `ts.factory.*` functions, rather than the (create|update|etc.)
functions that were originally exported directly on `ts`. said functions
were deprecated in v4.0 of typescript, and are slated to be removed in
v5.0.

this commit has two motivations.
1. these functions will be removed in typescript 5.0. this is a piece of
   technical debt the team was going to be tackling in preparation for
   said upgrade.
2. when attempting to upgrade to typescript 4.8, we ran into two issues:
   - several warnings regarding the use of these deprecated methods
     began to spring up, which wasn't a pleasant dev experience
   - it broke stencil - static members were not being properly
     transpiled from the original source code to our intermediate
     representations. code that was originally written as:
```tsx
@Component({ ... }) export class Foo {
  @Prop bar: boolean;
}
```
    would be transformed to:
```tsx
export class {}
```

the changes in this PR help with the 4.8 upgrade, in that moving to
functions that are no longer deprecated allow for the code to be
transpile properly. in TS 4.8, decorators are no longer a separate field
on `Node` derived classes, but are added to the `modifiers` field. this
change does _not_ take those changes directly into account. it only
moves to the non-deprecated factory functions. the upgrade to TS 4.8
will modify some of these functions calls to manage `Node` typing
changes