chore: bin files eslint rule override (#3794)
this commit adds an explicit override for the `bin` directory
and its usage of the "no-var" rule. the entry point to stencil
still uses the `var` keyword. until we change that, we ignore
errors found in that file.

a follow up pull request, #3797 will be added to
programmatically lint the `bin/` directory