chore(lint): lint scripts directory (#3803)
this commit updates the `lint` script in `package.json`
to lint the root level `scripts/` directory. because the
directory was not being linted previously, this commit
also resolves the lint violations as well.

certain function signatures needed to be updated to
more accurately reflect their return values as a part of
fixing lint violations. similarly, some function bodies were
updated to have an explicit return (rather than an implicit
one).