refactor(release): remove hidden compilation step (#3814)
this commit converts the entrypoint for our release scripts to use typescript
and more modern javascript-ecosystem paradigms. this is being done
as a first pass at modernizing our release scripts.

a consequence on converting the release entrypoint to typescript is that
the entrypoint no longer compiles the release scripts. to account for this,
the `package.json` scripts for running stencil builds and releases have been
updated to ensure that the scripts are compiled beforehand.

the output of the release scripts currently share `scripts/build` with other
build processes. as a result, it is possible for the typescript compiler to
erroneously think the project is up to date when we try to compile it and
skip compilation (e.g. see `sys-node-bundles-cache/`). the `--force` flag
is added here to ensure the scripts are compiled.