test(transform-utils): add isMemberPrivate unit tests (#3832)
this commit adds unit tests to the `isMemberPrivate` function in the
transform-utils module. this is being done in preparation for changes to
the function that are to occur in the typescript 4.8 upgrade. adding
unit tests will provide additional safety during the upgrade, but aren't
strictly necessary for it. the hope is to land this commit to the
primary branch prior to the upgrade.