test(convert-decorators): add unit test to filter util
this commit adds unit tests to the `filterDecorators` utility file. this
is done in an effort to make this utility a more maintainable function,
by taking the changes to remove strict null checks violations in 611756b
(#3597) and improve its testing.

adding tests to this function will aid in improving the confidence in
refactoring this function. the refactor will occur in a later commit,
but is planned to take place as a predecessor to the typescript 4.8
upgrade, as v4.8 changes how decorators are retrieved off of nodes in
the syntax tree, affecting this function.

the function is made public as a result of this commit. although it
breaks encapsulation of the module, it allows us to easily get this
function under test for the typescript 4.8 upgrade + necessary
refactors. we accept that in favor of better testing/higher confidence
in the refactor.