chore(release): re-enable custom release versions (#3849)
this commit updates the prompts for the release scripts to allow for
custom versions and tags to be specified. an issue was occurring where
two inputs with the same name would only allow the first to be run. this
commit updates the second version & tag prompts to have a unique name,
where the two values are coalesced after either question is asked.