chore(bundler-test): add readme, dependabot support (#3840)
this commit adds a readme to the bundler test directory to give a high
level overview what the directory is for. this commit also adds the
directory to dependabot configuration. the latter will allow us to get
ahead of any major breaking changes in the small dependencies list that
this project has.