refactor(filterDecorators): narrow first parameter  (#3835)
this commit narrows the first parameter type of `filterDecorators` to
accept a list of decorator nodes, rather than a single node itself and
subsequently pulling the decorators off the node as the first argument.

this change has two connected motivating factors:
1. this function can return either a new array of decorators, or the
   existing one. for cases where referential equality is important (i.e.
   comparing if the array provided to this function is the same as the
   one returned), it will be beneficial to have a reference to that
   array before calling this function because:
2. in typescript 4.8, decorators will no longer have direct accessors on
   a node. instead, helper functions provided by the language will be
   used to retrieve decorators. these helper methods allocate a new
   array when returning an array of decorators, making referential
   equality checks impossible as they are today (calling the
   `decorators` getter directly today will return a ref to the same
   array)

this commit adds unit tests to the `filterDecorators` utility fn. this
is done in an effort to make this utility a more maintainable function,
by taking the changes to remove strict null checks violations in https://github.com/ionic-team/stencil/commit/611756be93e4cd7a945fba0068c8502a35a1a90a
(https://github.com/ionic-team/stencil/pull/3597) and improve its testing.

the function is made public as a result of this commit. although it
breaks encapsulation of the module, it allows us to easily get this
function under test for the typescript 4.8 upgrade + necessary
refactors. we accept that in favor of better testing/higher confidence
in the refactor.