feat(typescript): add support for typescript 4.9
this commit increments typescript to v4.9 for the project, and updates
the dependabot configuration to not create a pull request for a version
of typescript newer than 4.9.

the `@types/node` package for the karma test suite is updated as a part
of this commit, as the currently installed version has problems being
parsed by v4.9. this is fixed in the type declarations themselves, hence
the upgrade.

the remainder of the changes in this pr are for fixing type errors where
there is no overlap between two types being compared. this occurs in two
scenarios:

  1. parsing css. in the css parser, there were three cases where an
     overlap did not occur in a comparison. in reading through the code,
     i cannot find a scenario where such a value could be returned such
     that the comparison is valid. as a result, such comparisons are
     removed.

     prior to removing these comparisons, an alternative (else block)
     was written with a logging statement in it. stencil's unit/karma
     tests were run, as well as the ionic framework's tests with this
     branch installed. the logging statement was never triggered. that
     in addition to the rich testing available for css, makes us
     confident that this change is safe to make.

  2. validation of a configuration flag. this occurred for the dev
     server, hydration config, and service worker config. these
     fields _should_ be properly set via type checking. if users are
     providing invalid values, we may want to reconsider some of the
     "silent failures" occurring and defaults we're setting in cases
     like this.