chore(dependabot): ignore @types/node versions for test/bundler (#3887)
this commit adds versions 17 & 18 of `@types/node` to the ignore list
for the `test/bundler` directory. we add these values because when we
run these tests locally & in a continuous integration env, we inherit
the version of node that is declared under the `volta` key in this
project's root level `package.json` file. to prevent version skew
either between the version of node and @types/node within bundler, we
need to add this property to the dependabot configuration.