test(app-types): update tests following jsdoc generation
this commit adds jsdoc to the output of the tests for app type
generation, following https://github.com/ionic-team/stencil/pull/3525
landing in `main`, then merging `main` into `v3.0.0-dev` caused tests
present in the v3 branch to begin to fail.

the alignment of the jsdoc being generated is off. in the name of time,
we assume tech-debt and will rectify this in stencil-660