chore(config): mark appendChildSlotFix as deprecated
This marks `extras.appendChildSlotFix` on the Stencil config as
deprecated, paving the way to remove the code for it later on. This also
adds TODO comments around the codebase nothing code which should be
removed when we remove the field entirely.

BREAKING CHANGES: The `appendChildSlotFix` field on `config.extras` is
deprecated, since support for the browsers which require this fix has
been dropped.