chore(config): mark `scriptDataOpts` field as deprecated
This marks the `extras.scriptDataOpts` field on the Stencil config as
deprecated and sprinkles TODO here and there to note where we'll need to
rip the related code out later.

BREAKING CHANGES: The `extras.scriptDataOpts` field is deprecated and
will be removed entirely in a future version. For now, it is renamed to
`__deprecated__scriptDataOpts`.